renderer: rewrite texture loader to get dimensions

+++ utilities/common/ptExtent.hpp
@@ -4,8 +4,8 @@
 
 struct PtExtent
 {
-	uint64_t height;
-	uint64_t width;
+	uint32_t height;
+	uint32_t width;
 };
 
 namespace platypus
@@ -13,8 +13,9 @@ namespace platypus
 
 struct Extent
 {
-	uint64_t height;
-	uint64_t width;
+	// dx uses UINTs which are only 32bits long
+	uint32_t height;
+	uint32_t width;
 };
 
 };
